requirements.txt
pandas
numpy
psycopg2-binary
sqlalchemy
streamlit
plotly
python-dotenv
